Browsers are terrible
Chrome (and maybe other browsers) do not send the `iframe` loaded event
unless you assign the src attribute. However, using `about:blank` as the
src causes issues on some browsers when serving over https and data uris
are considered cross-origin and can not be loaded into a sandboxed iframe.

The only solution is to have a blank html page on the server that we
request when initializing the frame.